LayoutTests/editing/pasteboard/interchange-newline-3-expected.txt

EDITING DELEGATE: shouldBeginEditingInDOMRange:range from 0 of DIV > BODY > HTML > #document to 0 of DIV > BODY > HTML > #document
EDITING DELEGATE: webViewDidBeginEditing:WebViewDidBeginEditingNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 0 of DIV > BODY > HTML > #document to 0 of DIV > BODY > HTML > #document toDOMRange:range from 0 of DIV > DIV > BODY > HTML > #document to 0 of DIV > DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
EDITING DELEGATE: webViewDidEndEditing:WebViewDidEndEditingNotification
This demonstrates a bug in interchange newline handling during paste.
There is an interchange newline at the end of the incoming fragment and so the caret should end up on an empty line.
The editable region is followed by non-editable content, which breaks the old code in paste that handled interchange newlines.
| "
"
| <div>
|   contenteditable="true"
|   id="test"
|   "foo"
|   <div>
|     "bar"
|   <div>
|     <#selection-caret>
|     <br>
| "
This is non-editable content.

"
| <script>
|   "
runTest();
"
